Replace linter with jsonschema validation using the SHARE schema, now all harvesters fail validation
