+++ requirements.txt
@@ -13,3 +13,4 @@ elasticsearch==1.3.0
 Flask==0.10.1
 blist==1.3.6
 furl==0.4.4
+jsonschema==2.4.0
