+++ scrapi/__init__.py
@@ -1,2 +1,1 @@
 from scrapi import harvesters  # noqa
-from scrapi.base import registry  # noqa
