+++ scrapi/base/__init__.py
@@ -7,11 +7,10 @@ import logging
 from datetime import date, timedelta
 
 from lxml import etree
-from celery.schedules import crontab
 
 from scrapi import util
 from scrapi import requests
-from scrapi.linter import lint
+from scrapi import registry
 from scrapi.base.schemas import OAISCHEMA
 from scrapi.base.helpers import updated_schema
 from scrapi.linter.document import RawDocument, NormalizedDocument
@@ -20,32 +19,6 @@ from scrapi.base.transformer import XMLTransformer, JSONTransformer
 logging.basicConfig(level=logging.INFO)
 
 logger = logging.getLogger(__name__)
-
-
-class _Registry(dict):
-
-    def __init__(self):
-        super(_Registry, self).__init__()
-
-    def __getitem__(self, key):
-        try:
-            return super(_Registry, self).__getitem__(key)
-        except KeyError:
-            raise KeyError('No harvester named "{}"'.format(key))
-
-    @property
-    def beat_schedule(self):
-        return {
-            'run_{}'.format(name): {
-                'args': [name],
-                'schedule': crontab(**inst.run_at),
-                'task': 'scrapi.tasks.run_harvester',
-            }
-            for name, inst
-            in self.items()
-        }
-
-registry = _Registry()
 
 
 class HarvesterMeta(abc.ABCMeta):
@@ -88,9 +61,6 @@ class BaseHarvester(object):
     @abc.abstractmethod
     def normalize(self, raw_doc):
         raise NotImplementedError
-
-    def lint(self):
-        return lint(self.harvest, self.normalize)
 
     @property
     def run_at(self):
